Fix migration to safely remove class_code column

+++ db/migrate/20251112043033_remove_class_code_from_users.rb
@@ -1,5 +1,5 @@
 class RemoveClassCodeFromUsers < ActiveRecord::Migration[7.1]
   def change
-    remove_column :users, :class_code, :string
+    remove_column :users, :class_code, :string if column_exists?(:users, :class_code)
   end
 end
